cancer-db/requirements.txt
Django==1.8.4
Parsley==1.3
PyYAML==3.11
XlsxWriter==0.7.3
awsebcli==3.5
backports.ssl-match-hostname==3.4.0.2
blessed==1.9.5
botocore==1.1.7
cement==2.4.0
colorama==0.3.3
diff-match-patch==20121119
django-import-export==0.2.8
django-parsley==0.6
django-selectable==0.9.0
django-session-security==2.2.4
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.7.1
pathspec==0.3.3
python-dateutil==2.4.2
requests==2.6.2
six==1.9.0
tablib==0.10.0
texttable==0.8.3
wcwidth==0.1.4
websocket-client==0.32.0
wsgiref==0.1.2
